requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.3 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
async-generator==1.10 ; python_version >= '3.5'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.1.0 ; python_version >= '3.5'
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3 ; python_version >= '3.7'
cryptography==38.0.3 ; python_version >= '3.6'
exceptiongroup==1.0.1 ; python_version < '3.11'
flask==2.2.2
frozenlist==1.3.3 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
idna==3.4
importlib-metadata==5.0.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
multidict==6.0.2 ; python_version >= '3.7'
outcome==1.2.0 ; python_version >= '3.7'
oyaml==1.0
pillow==9.3.0
psutil==5.9.4
pycparser==2.21
pyopenssl==22.1.0
pysocks==1.7.1
pyyaml==6.0
qrcode==7.3.1
requests==2.28.1
selenium==4.6.0
slack-sdk==3.19.3
sniffio==1.3.0 ; python_version >= '3.7'
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0 ; python_version >= '3.7'
trio-websocket==0.9.2 ; python_version >= '3.5'
urllib3==1.26.12
werkzeug==2.2.2 ; python_version >= '3.7'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
yarl==1.8.1 ; python_version >= '3.7'
zipp==3.10.0 ; python_version >= '3.7'
